Linter autofix

+++ pioneer/packages/joy-utils/src/transport/validators.ts
@@ -3,7 +3,8 @@ import { u32 } from '@polkadot/types/';
 
 export default class ValidatorsTransport extends BaseTransport {
   async maxCount (): Promise<number> {
-    const count = ((await this.api.query.staking.validatorCount()) as u32).toNumber();
+    const count = ((await this.api.query.staking.validatorCount())).toNumber();
+
     return count;
   }
 }
